requirements/base.txt
# Django
Django==1.11.6

# Environmental var configuration
django-configurations==2.0

# For the persistence stores
psycopg2==2.7.3.2
dj-database-url==0.4.2

# Models
django-model-utils==3.0.0

# Rest apis
djangorestframework==3.7.1
Markdown==2.6.9
django-filter==1.1.0

# Static files
whitenoise==3.3.1

# Media files
boto==2.48.0
django_unique_upload==0.2.1

# Image manipulation
django-versatileimagefield==1.8.1

# For asynchronous queuing
django-rq==0.9.6

# Time zones support
pytz==2017.3

# Query inspection
django-queryinspect==1.0.0

# CORS Haders
django-cors-headers==2.1.0

# Requests
chardet==3.0.4
urllib3==1.22
requests==2.18.4

# Celery
six==1.10.0
amqp==1.4.9
billiard==3.3.0.23
celery==3.1.23
kombu==3.0.37

# Opbeat
opbeat==3.4

# Braintree
braintree==3.39.0